优化箱线图展示，可以自定义5个名字

+++ NewLife.CubeNC/Charts/Models/BoxplotItem.cs
@@ -6,4 +6,4 @@
 /// <param name="Median">中位数</param>
 /// <param name="Q3">上四分位数</param>
 /// <param name="Max">最大值</param>
-public record BoxplotItem(Single Min, Single Q1, Single Median, Single Q3, Single Max);
+public record BoxplotItem(Double Min, Double Q1, Double Median, Double Q3, Double Max);
